+++ Sources/ZcashLightClientKit/Error/ZcashErrorCode.swift
@@ -255,8 +255,8 @@ public enum ZcashErrorCode: String {
     case rustFetchUTXOsByAddress = "ZRUST0089"
     /// Error from rust layer when calling ZcashRustBackend.deleteAccount
     case rustDeleteAccount = "ZRUST0090"
-    /// Error from rust layer when calling ZcashRustBackend.rewindToChainState
-    case rustRewindToChainState = "ZRUST0091"
+    /// Error from rust layer when calling ZcashRustBackend.truncateToChainState
+    case rustTruncateToChainState = "ZRUST0091"
     /// Error from rust layer when calling ZcashEip681Backend.parseTransactionRequest
     case rustEip681Parse = "ZRUST0092"
     /// SQLite query failed when fetching all accounts from the database.